+++ quantforge/__init__.py
@@ -667,6 +667,14 @@ from .dag_paths import dag_shortest_path, dag_longest_path, transitive_closure
 from .min_cost_flow import MinCostMaxFlow, min_cost_max_flow
 from .fd_weights import fd_weights
 from .multistep_ode import adams_bashforth_moulton
+from .rotation import (
+    quat_to_matrix,
+    matrix_to_quat,
+    euler_to_quat,
+    quat_to_euler,
+    euler_to_matrix,
+    matrix_to_euler,
+)
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1371,7 +1379,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "4.56.0"
+__version__ = "4.57.0"
 
 __all__ = [
     "OptionType",
@@ -2288,6 +2296,12 @@ __all__ = [
     "min_cost_max_flow",
     "fd_weights",
     "adams_bashforth_moulton",
+    "quat_to_matrix",
+    "matrix_to_quat",
+    "euler_to_quat",
+    "quat_to_euler",
+    "euler_to_matrix",
+    "matrix_to_euler",
     "Dual",
     "dual_derivative",
     "dual_gradient",